requirements.txt
# Core Streamlit
streamlit==1.28.1

# AI/ML Libraries
openai==1.3.8
google-generativeai==0.3.2
transformers
torch==2.2.0
torchaudio==2.2.0
httpx==0.25.2
# Audio/Video Processing
yt-dlp==2025.05.22
openai-whisper==20231117
faster-whisper
gTTS==2.5.0
pydub>=0.25.1
ffmpeg-python>=0.2.0
pytube
google-cloud-speech>=2.21.0

# Japanese Language Processing
fugashi==1.3.0
unidic-lite==1.0.8
sudachipy
sudachidict_core
jaconv>=0.3.4

# Translation
deep-translator==1.11.4

# Data Processing
pandas==2.2.0
numpy>=1.24.0,<2.0.0
scipy

# Flashcard Generation
genanki==0.13.0

# Utilities
python-dotenv==1.0.0
langdetect>=1.0.9
aiohttp>=3.9.0
urllib3

# Additional utilities
Pillow>=10.0.0
python-multipart>=0.0.6

# Additional dependencies for compatibility
typing-extensions>=4.8.0
pydantic>=2.5.0
certifi>=2023.11.17
charset-normalizer>=3.3.2
idna>=3.6
requests>=2.31.0